requirements.txt
pytesseract
opencv-python
numpy
mysql-connector-python
pdf2image
pymupdf
python-multipart
